feat: dashboard container
