+++ src/Main.js
@@ -5,7 +5,7 @@ import { VerifyLinkTokenModal } from "./Components/VerifyLinkToken/VerifyLinkTok
 import { RegisterModal } from "./Components/Register/RegisterModal";
 import { AddressUpdateModal } from "./Components/AddressUpdate/AddressUpdateModal";
 import { SelectModalWithHook as SelectModal } from "./Components/Select/SelectModal";
-import { DashboardWithHook as Dashboard } from "./Components/dashboard/Dashboard";
+import { Dashboard } from "./Components/dashboard/Dashboard";
 import { UserUpdateModal } from "./Components/UserUpdate/UserUpdateModal";
 import { ProfilePicChangeModal } from "./Components/ProfilePicChange/ProfilePicChangeModal";
 import { PaymentMethodUpdateModal } from "./Components/PaymentMethodUpdate/PaymentMethodUpdateModal";
