+++ src/utils/action-types.js
@@ -114,3 +114,13 @@ export const HANDLE_REMOVE_MEMBER = "HANDLE_REMOVE_MEMBER";
 export const UPDATE_INVITE_BUTTON = "UPDATE_INVITE_BUTTON";
 export const UPDATE_MEMBERS = "UPDATE_MEMBERS";
 export const UPDATE_REMOVE_MEMBER_ID = "UPDATE_REMOVE_MEMBER_ID";
+//Dashboard
+export const SET_IS_OPEN = "SET_IS_OPEN";
+export const SET_ADDRESSES = "SET_ADDRESSES";
+export const CLOSE_DASHBOARD = "CLOSE_DASHBOARD";
+export const SET_ACTIVE_DASHBOARD_LINK = "SET_ACTIVE_DASHBOARD_LINK";
+export const SET_DISABLESUBMIT = "SET_DISABLESUBMIT";
+export const CANCEL_SUBSCRIPTION = "CANCEL_SUBSCRIPTION";
+export const UNSUSPEND_SUBSCRIPTION = "UNSUSPEND_SUBSCRIPTION";
+export const REACTIVATE_SUBSCRIPTION = "REACTIVATE_SUBSCRIPTION";
+//========
